+++ src/gpre/languages/ada.cpp
@@ -902,8 +902,8 @@ static void gen_blob_open( const act* action, USHORT column)
 
 static void gen_blr(void* /*user_arg*/, SSHORT /*offset*/, const char* string)
 {
-	const int c_len = strlen(COMMENT);
-	const int len = strlen(string);
+	const int c_len = static_cast<int>(strlen(COMMENT));
+	const int len = static_cast<int>(strlen(string));
 	int from = 0;
 	int to = 120 - c_len;
