+++ src/gpre/msc.cpp
@@ -445,7 +445,7 @@ gpre_req* MSC_request(req_t type)
 
 SCHAR* MSC_string(const TEXT* input)
 {
-	TEXT* string = (TEXT*) MSC_alloc(strlen(input) + 1);
+	TEXT* string = (TEXT*) MSC_alloc(static_cast<int>(strlen(input) + 1));
 	strcpy(string, input);
 
 	return string;
@@ -494,7 +494,7 @@ gpre_nod* MSC_unary(nod_t type, gpre_nod* arg)
 gpre_usn* MSC_username(const SCHAR* name, USHORT name_dyn)
 {
 	gpre_usn* username = (gpre_usn*) MSC_alloc(USN_LEN);
-	char* newname = (char*) MSC_alloc(strlen(name) + 1);
+	char* newname = (char*) MSC_alloc(static_cast<int>(strlen(name) + 1));
 	username->usn_name = newname;
 	strcpy(newname, name);
 	username->usn_dyn = name_dyn;
