+++ src/lock/print.cpp
@@ -1270,7 +1270,7 @@ static void prt_owner(OUTFILE outfile,
 	FPRINTF(outfile, "\tProcess id: %6d (%s), thread id: %6"SIZEFORMAT"\n",
 			process->prc_process_id,
 			ISC_check_process_existence(process->prc_process_id) ? "Alive" : "Dead",
-			owner->own_thread_id);
+			static_cast<size_t>(owner->own_thread_id));
 
 	const USHORT flags = owner->own_flags;
 	FPRINTF(outfile, "\tFlags: 0x%02X ", flags);
